backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.5
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
apscheduler==3.10.4
pytesseract==0.3.13
Pillow==10.4.0
google-api-python-client==2.142.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
